fix bug

+++ server/service/user.js
@@ -113,12 +113,14 @@ export const collectMovie = async (userId, movieId) => {
 export const modifyUser = async (data) => {
   const { id, email, username, password, job, headImg, birthday} = data
   let user = await User.findOne({_id: id})
-  fs.unlinkSync(path.resolve(__dirname, `../../public/${user.headImg}`))
+  if (headImg){
+    fs.unlinkSync(path.resolve(__dirname, `../../public/${user.headImg}`))
+    user.headImg = headImg
+  }
   email && (user.email = email) 
   username && (user.username = username)
   password && (user.password = password)
   job && (user.job = job)
-  headImg && (user.headImg = headImg)
   birthday && (user.birthday = birthday)
   try {
     await user.save()
